Bring default manifest to 1.7 release.
The current default manifest has two issues.  One is that it uses
git@github.com style remotes that then generate complaint messages
and the other is that we'd inadvertently let the branches roll
forward, picking up touch command dependencies.

This fixes it by locking it down to specific, tested, working revs
and sets the remotes to https style.